+++ src/exec/packer/mod.rs
@@ -1,6 +1,7 @@
 // region:    --- Modules
 
 mod pack_toml;
+mod provenance;
 mod support;
 
 mod installer_impl;